Clifford Wolf: 	Fixed two compiler warnings
git-svn-id: http://svn.clifford.at/openscad/trunk@51 b57f626f-c46c-0410-a088-ec61d464b74c

+++ expr.cc
@@ -73,7 +73,7 @@ Value Expression::evaluate(const Context *context) const
 		Value v1 = children[0]->evaluate(context);
 		Value v2 = children[1]->evaluate(context);
 		if (v1.type == Value::VECTOR && v2.type == Value::NUMBER) {
-			int i = v2.num;
+			int i = (int)(v2.num);
 			if (i < v1.vec.size())
 				return *v1.vec[i];
 		}
